requirements.txt
python-dotenv
transformers
torch 
torchvision 
torchaudio
huggingface_hub
langchain
openai
streamlit
watchdog